@@ -18,7 +18,7 @@ typedef void* HANDLE;
 
 #include "DynamsoftCore.h"
 
-#define DBR_VERSION "11.0.60.5318"
+#define DBR_VERSION "11.2.10.5856"
 
 /**Enumeration section*/
 
@@ -226,6 +226,10 @@ typedef enum LocalizationMode
 
 	/**Localizes 1D barcodes fast. Check @ref LM for available argument settings. */
 	LM_ONED_FAST_SCAN = 0x100,
+
+	// added in v11.0.10
+	/** Localizes barcodes by utilizing a neural network model. */
+	LM_NEURAL_NETWORK = 0X200,
 
 	/**Reserved setting for localization mode.*/
 #if defined(_WIN32) || defined(_WIN64)
@@ -1346,7 +1350,6 @@ namespace dynamsoft
 				 */
 				virtual int SetDecodedBarcode(const CDecodedBarcodeElement* element, const double matrixToOriginalImage[9] = IDENTITY_MATRIX) = 0;
 
-
 				virtual int RemoveDecodedBarcode(int index) = 0;
 			};
 		}
@@ -1468,6 +1471,7 @@ namespace dynamsoft
 			 *
 			 */
 			virtual int SetLocation(const CQuadrilateral& location) = 0;
+
 		};
 
 		/**
