@@ -27,7 +27,7 @@
 #include "DynamsoftCodeParser.h"
 //#endif
 
-#define DCV_VERSION                  "3.0.60.5318"
+#define DCV_VERSION                  "3.2.10.5856"
 
 /**Enumeration section*/
 
@@ -698,6 +698,10 @@ namespace dynamsoft
 			*
 			*/
 			virtual CImageData* GetOriginalImage(const char* imageHashId) = 0;
+
+			virtual void AddOriginalImageToCache(const char* imageHashId) = 0;
+
+			virtual void RemoveOriginalImageFromCache(const char* imageHashId) = 0;
 		};
 
 		/**
@@ -794,6 +798,7 @@ namespace dynamsoft
 
 		};
 
+		class CVR_API CCaptureVisionRouter;
 		/**
 		* The `CCapturedResultFilter` class is responsible for filtering captured results. It contains several callback functions for different types of results, including original image, decoded barcodes, recognized text lines, detected quads, deskewed images, and parsed results.
 		*/
@@ -884,7 +889,7 @@ namespace dynamsoft
 			* Initializes the filter. It will be called by Capture Vision Router before using the filter.
 			*
 			*/
-			virtual void Init();
+			virtual void Init(CCaptureVisionRouter* router = nullptr);
 
 			virtual const char* GetEncryptedString();
 
@@ -1061,7 +1066,7 @@ namespace dynamsoft
 			* Loads and initializes a template from a string.
 			*
 			* @param [in] content The string containing the template.
-			* @param [in] errorMsgBuffer A buffer for error messages.
+			* @param [in,out] errorMsgBuffer A buffer for error messages.
 			* @param [in] errorMsgBufferLen The length of the error message buffer.
 			*
 			* @return Returns an error code. Zero indicates success.
@@ -1073,7 +1078,7 @@ namespace dynamsoft
 			* Loads and initializes a template from a file.
 			*
 			* @param [in] filePath The path to the file containing the template.
-			* @param [in] errorMsgBuffer A buffer for error messages.
+			* @param [in,out] errorMsgBuffer A buffer for error messages.
 			* @param [in] errorMsgBufferLen The length of the error message buffer.
 			*
 			* @return Returns an error code. Zero indicates success.
@@ -1121,7 +1126,7 @@ namespace dynamsoft
 			*
 			* @param [in] templateName The name of the template to update.
 			* @param [in] settings A pointer to a `SimplifiedCaptureVisionSettings` object.
-			* @param [in] errorMsgBuffer A buffer for error messages.
+			* @param [in,out] errorMsgBuffer A buffer for error messages.
 			* @param [in] errorMsgBufferLen The length of the error message buffer.
 			*
 			* @return Returns an error code. Zero indicates success.
@@ -1297,7 +1302,7 @@ namespace dynamsoft
 			*
 			* @param [in] waitForThreadExit Indicates whether to wait for the capture process to complete before returning. The default value is false.
 			*
-			* @param [out] errorMsgBuffer Stores any error messages generated during the capturing process. If no buffer is provided, the error messages will not be output.
+			* @param [in,out] errorMsgBuffer Stores any error messages generated during the capturing process. If no buffer is provided, the error messages will not be output.
 			*
 			* @param [in] errorMsgBufferLen Specifies the length of the provided error message buffer. If no buffer is provided, this parameter is ignored.
 			*
@@ -1368,6 +1373,26 @@ namespace dynamsoft
 			* @return Returns 0 if succeeds, nonzero otherwise.
 			*/
 			static int AppendModelBuffer(const char* modelName, const unsigned char* modelBytes, int modelBytesLength, int maxModelInstances);
+
+			/**
+			* Switch the capturing template during the image processing workflow.
+			*
+			* @param [in] templateName The name of the new capturing template to apply.
+			* @param [in,out] errorMsgBuffer A buffer for error messages.
+			* @param [in] errorMsgBufferLen The length of the error message buffer.
+			*
+			* @return Returns an error code. Zero indicates success.
+			*/
+			int SwitchCapturingTemplate(const char* templateName, char errorMsgBuffer[] = NULL, const int errorMsgBufferLen = 0);
+
+			/*
+			* Sets the global number of threads used internally for model execution.
+			*
+			* @param [in] intraOpNumThreads Number of threads used internally for model execution. Valid range: [0, 256]
+			*                               If the value is outside the range [0, 256], it will be treated as 0 (default).
+			*/
+			static void SetGlobalIntraOpNumThreads(int intraOpNumThreads = 0);
+
 		private:
 			CCaptureVisionRouter(const CCaptureVisionRouter& r);
 			CCaptureVisionRouter& operator=(const CCaptureVisionRouter& r);
