@@ -13,7 +13,7 @@
 #endif
 #include "DynamsoftCore.h"
 
-#define DIP_VERSION                  "3.0.60.5318"
+#define DIP_VERSION                  "3.2.10.5856"
 
 namespace dynamsoft
 {
